refactored tool outputs
