+++ templates/ToolPage/Outputs/QuizResponse/QuizResponse.jsx
