+++ templates/ToolPage/Outputs/QuizResponse/index.js
