+++ templates/ToolPage/Outputs/QuizResponse/styles.js
