+++ templates/ToolPage/Outputs/SyllabusGeneratorResponse/SyllabusGeneratorResponse.jsx
