+++ templates/ToolPage/Outputs/SyllabusGeneratorResponse/index.js
