+++ templates/ToolPage/Outputs/SyllabusGeneratorResponse/styles.js
